refactor(scripts): combine OG image generation scripts
Merge generate-projects-og.js into generate-page-og.js to create a
single unified script for all page OG images (home, blog, categories,
projects). Run with `npm run cover:page-og [page]` or `npm run
cover:page-og all` to generate specific or all page images.
