+++ src/pages/blog/[...page].astro
@@ -15,7 +15,7 @@ export const getStaticPaths = (async ({ paginate }) => {
 const { page } = Astro.props;
 ---
 
-<BaseLayout title="Blog" description="All blog posts from Coding With Calvin">
+<BaseLayout title="Blog" description="All blog posts from Coding With Calvin" image="/images/blog-og.png">
   <section class="py-12">
     <div class="container mx-auto px-4">
       <h1 class="text-4xl font-heading mb-8">Blog</h1>
