+++ src/pages/categories.astro
@@ -19,7 +19,7 @@ const categories = Array.from(categoryMap.entries()).sort((a, b) =>
 );
 ---
 
-<BaseLayout title="Categories" description="Browse posts by category">
+<BaseLayout title="Categories" description="Browse posts by category" image="/images/categories-og.png">
   <section class="py-12">
     <div class="container mx-auto px-4">
       <h1 class="text-4xl font-heading mb-8">Categories</h1>
