+++ src/pages/index.astro
@@ -9,7 +9,7 @@ const latestPosts = allPosts
   .slice(0, 6);
 ---
 
-<BaseLayout title="Who am I?" description="Software development blog by Calvin Allen">
+<BaseLayout title="Who am I?" description="Software development blog by Calvin Allen" image="/images/home-og.png">
   <section class="py-12">
     <div class="container mx-auto px-4">
       <div class="relative">
